+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## next
 ### Added
 - Added `BuildFeature` feature flags. See [Feature Flags](https://github.com/JetBrains/gradle-intellij-plugin#build-features) in README.
+- Added `--jbr-illegal-access` and `-XX:+IgnoreUnrecognizedVMOptions` flags for tasks based on `RunIdeBase` to support `2022.2` which runs on Java 17
 - `instrumentCode` task – incremental instrumentation [#459](../../issues/459)
 - Add `intellijRepository` to the beginning of the repositories list when resolving dependencies [#615](../../issues/615)
 - Set `-Djdk.module.illegalAccess.silent=true` flag by default to mute `WARNING: An illegal reflective access operation has occurred`